Updates in pages
