+++ src/main/webapp/index.html
@@ -1,41 +1,205 @@
-<html lang="en"><head>
+<!DOCTYPE html>
+<html lang="en">
+<head>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
   <meta name="description" content="">
   <meta name="author" content="">
+  <link rel="stylesheet" href="https://use.typekit.net/nsf1sey.css">
 
   <title>HOME PAGE | FWRP</title>
 
-  <link rel="canonical" href="https://getbootstrap.com/docs/4.0/examples/cover/">
+  <style>
+    body {
+      font-family: Arial, sans-serif;
+      margin: 0;
+      padding: 0;
+      display: flex;
+      flex-direction: column;
+      min-height: 100vh;
+      background-color: #ededed;
+    }
 
-  <!-- Bootstrap core CSS -->
-  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.4.1/dist/css/bootstrap.min.css" integrity="sha384-Vkoo8x4CGsO3+Hhxv8T/Q5PaXtkKtu6ug5TOeNV6gBiFeWPGFN9MuhOf23Q9Ifjh" crossorigin="anonymous">
+    .header {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      background-color: black;
+      /*border-radius: 25px;*/
+      padding: 11px 22px;
+      color: black;
+      position: fixed;
+      left: 50%;
+      transform: translateX(-50%);
+      height: 5vh;
+      width: 95%;
+      background: rgba(30, 23, 23, 0.1);
+      backdrop-filter: blur(50px);
+      border-radius: 30px;
+      z-index: 1000;
+    }
 
-  </head>
+    .header h1 {
+      font-size: 1.5em;
+      margin: 0;
+      color: black;
+    }
 
-<body class="text-center">
+    .header a {
+      text-decoration: none;
+      background-color: #FFBF00;
+      color: black;
+      padding: 10px 25px;
+      border-radius: 25px;
+      transition: background-color 0.3s ease;
+      font-size: 14px;
+      margin-left: 10px;
+    }
 
-  <div class="cover-container d-flex h-100 p-3 mx-auto flex-column">
-    <header class="masthead mb-auto">
-      <div class="inner">
-        <nav class="nav nav-masthead justify-content-center">
-          <a class="nav-link active" href="./index.html">Home</a>
-          <a class="nav-link" href="./login.jsp">Login</a>
-          <a class="nav-link" href="./register.jsp">Register</a>
-        </nav>
-      </div>
-    </header>
+    .header a:hover {
+      background-color: black;
+      color: white;
+    }
 
-    <main role="main" class="inner cover">
-      <h1 class="cover-heading">Welcome to our Food Waste Reduction Platform</h1>
-       
-    </main>
+    .main-content {
+      margin-top: 50px; 
+      position: relative;
+      text-align: center;
+    }
+    
+    .herotxt {
+        font-family: owners;
+        font-weight: 800px;
+      font-size: 7vh;
+      position: absolute;
+      top: 50%;
+      left: 50%;
+      transform: translate(-50%, -50%);
+      color: white;
+      font-weight: 500px;
+    }
+    
+    .herotxt .explore-button {
+      display: inline-block;
+      text-decoration: none;
+      background-color: #FFBF00;
+      color: black;
+      padding: 10px 25px;
+      border-radius: 25px;
+      transition: background-color 0.3s ease;
+      font-size: 14px;
+      margin-top: 20px;
+    }
+    
+    .herotxt .explore-button:hover {
+      background-color: black;
+      color: white;
+    }
+    
+    .video-container {
+      margin: 10px;
+      height: 400px;
+      position: relative;
+      overflow: hidden;
+      border-radius: 10px;
+    }
+    
+    video {
+      width: 100%;
+      height: auto;
+    }
+
+    .section {
+        padding: 15px;
+        font-family: owners;
+        display: flex;
+        flex-direction: row;
+        gap:10px;
+    }
+    
+    .left{
+        font-weight: 100px;
+        background-color: #BDCD4B;
+        width:50%;
+        padding: 50px;
+        border-radius: 50px;
+        height:450px;
+    }
+    .left1{
+        font-size: 5vh;
+    }
+    .right{
+        background-color: #f7cb43;    
+        width:50%;
+        padding: 50px;
+        border-radius: 50px;
+        height:450px;
+    }
+    .right1{
+        font-size: 5vh;
+    }
+    
+    .sec2left{
+        font-family: owners;
+        color: white;
+        font-size: 4vh;
+        display: flex;
+        height:400px;
+        padding: 20px;
+        gap: 10px;
+        margin: 15px;
+        background-color: black;
+        /*background-color: #C8B6FF;*/
+        border-radius: 50px;
+    }
+    .sec2left{
+        font-size:4.5vh;
+    }
+  </style>
+</head>
+<body>
+  <div class="header">
+    <h1>FWRP</h1>
+        <div class="header-buttons">
+      <a href="LoginServlet" class="button">Login</a>
+    </div>
   </div>
 
+  <div class="main-content">
+    <div class="video-container">
+      <video autoplay loop muted>
+        <source src="res/hero.mp4" type="video/mp4">
+      </video>
+      <h1 class="herotxt">
+        FOOD WASTE REDUCTION PLATFORM <br>
+        <a href="#section1" class="explore-button">Explore</a>
+      </h1>
+    </div>
+  </div>
 
-  <script src="https://code.jquery.com/jquery-3.4.1.slim.min.js" integrity="sha384-J6qa4849blE2+poT4WnyKhv5vZF5SrPo0iEjwBvKU7imGFAV0wwj1yYfoRSJoZ+n" crossorigin="anonymous"></script>
-  <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js" integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo" crossorigin="anonymous"></script>
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.4.1/dist/js/bootstrap.min.js" integrity="sha384-wfSDF2E50Y2D1uUdj0O3uMBJnjuUD4Ih7YwaYd1iqfktj0Uod8GCExl3Og8ifwB6" crossorigin="anonymous"></script>
+  <div id="section1" class="section">
+      <div class="left"><h1>Join Us in Making a Difference</h1>
+          <p class="left1">Be a part of our mission to reduce food waste and promote sustainability. Explore how you can contribute, whether you're a retailer, a consumer, or a charitable organization. Together, we can create a more efficient and equitable food distribution system.</p>
+      </div>
+            
+      
+      <div class="right"><h1>Our vision.</h1>
+          <p class="right1">At FWRP, we envision a world where surplus food is transformed into opportunities rather than waste. By fostering collaboration among stakeholders, we aim to create a sustainable ecosystem that supports both people and the planet.</p>
+      </div>
 
+  </div>
 
-</body></html>
+    <div id="section2">
+  <div class="sec2left">
+    <ul>
+      <li>Efficient User Registration: Seamlessly create accounts for retailers, consumers, or charitable organizations.</li>
+      <li>Comprehensive Inventory Management: Retailers can easily manage and list surplus food items.</li>
+      <li>Consumer Access to Discounts: Browse and purchase surplus food at reduced prices with notifications on new offers.</li>
+      <li>Charitable Organizations: Claim surplus food from retailers to help those in need.</li>
+      <li>Real-Time Surplus Food Alerts: Receive instant notifications about available surplus food based on preferences and location.</li>
+    </ul>
+  </div>
+</div>
+
+</body>
+</html>
